+++ src/components/Schedule/AdditionalContent.css
@@ -1,5 +1,4 @@
 .AdditionalContent{
-    position: relative;
     display: flex;
     flex-direction: column;
