+++ src/components/Schedule/BladeTaskCardOverlay.tsx
@@ -6,32 +6,20 @@ import { BladeTaskCardProps } from "./BladeTaskCard";
 
 function BladeTaskCardOverlay(props: BladeTaskCardProps) {
 
-
     let cardStyle: React.CSSProperties;
     //Dynamic styling based on props values
-    if (props.startDate) {
-        cardStyle = {
-            backgroundColor: props.shown ? props.projectColor : "grey",
-            gridColumn: `date-${props.startDate.getFullYear()}-${props.startDate.getMonth()}-${props.startDate.getDate()} / span ${
-                props.duration
-            }`,
-            border: props.inConflict ? "2px dashed red" : "",
-        };} else{
-            cardStyle = {
-                backgroundColor: props.shown ? props.projectColor : "grey",
-                width: `${props.duration * dateDivLength}px`,
-                border: props.inConflict ? "2px dashed red" : "",
-                gridRow: `project-${props.projectName}`,
-                gridColumn: "2",
-                justifyContent: "left",
-            };
-        }
+    cardStyle = {
+        backgroundColor: props.shown ? props.projectColor : "grey",
+        width: `${props.duration * dateDivLength}px`,
+        border: props.inConflict ? "2px dashed red" : "",
+    };
+
 
         return (
             <div
             className="bladeTaskCard"
             style={cardStyle}
-            id={`${props.id} Overlay`}
+            id={`Overlay-id-${props.id}`}
         >
             {/* {used to visualize the attach period} */}
             <div
