+++ src/components/Schedule/Display.tsx
@@ -3,19 +3,20 @@ import CreateTestRigDivs from "./TestRigDivs";
 import CreateTimelineField from "./TimelineField";
 import React, { useEffect, useState } from "react";
 import CreateAdditionalContent from "./AdditionalContent";
-import BladeTaskCard from "./BladeTaskCard";
-import { useQuery, useSubscription } from "@apollo/client";
-import {
-    GET_BT_IN_RANGE_SUB,
-    GET_TEST_RIGS,
-    GET_BT_PENDING_SUB,
-    GET_ALL_BLADE_PROJECTS,
-} from "../../api/queryList";
+import BladeTaskCard, { BladeTaskCardProps } from "./BladeTaskCard";
+import { useMutation, useQuery, useSubscription } from "@apollo/client";
+import { GET_BT_IN_RANGE_SUB, GET_TEST_RIGS, GET_BT_PENDING_SUB, GET_ALL_BLADE_PROJECTS } from "../../api/queryList";
 import { getMonthLength } from "./TimelineField";
 import { capitalizeFirstLetter } from "./TimelineField";
 import { useEditModeContext } from "../../EditModeContext";
 import SwitchComponent from "../TableLogic/SwitchComponent";
 import StyledButton from "../ui/styledButton";
+import PendingTasks from "./PendingTasks";
+import { DndContext, DragOverlay } from "@dnd-kit/core";
+import BladeTaskCardOverlay from "./BladeTaskCardOverlay";
+import { createPortal } from "react-dom";
+import { BladeTaskHolder } from "./BladeTaskHolder";
+import { UPDATE_BT } from "../../api/mutationList";
 
 const currentDate = new Date(Date.now()); // Get the current date
 
@@ -28,43 +29,37 @@ type DisplayProps = {
 
 function DisplayComponent(props: DisplayProps) {
     const editMode = useEditModeContext();
-    const [rigs, setRigs] = useState<{ rigName: string; rigNumber: number }[]>([
-        { rigName: "No Rigs", rigNumber: 0 },
-    ]);
+    const [rigs, setRigs] = useState<{ rigName: string; rigNumber: number }[]>([{ rigName: "No Rigs", rigNumber: 0 }]);
 
-    const [selectedDate, setSelectedDate] = useState(
-        `${currentDate.getFullYear()}-${
-            currentDate.getMonth() + 1
-        }-${currentDate.getDate()}`
-    ); // State to store the selected date
+    const [selectedDate, setSelectedDate] = useState(`${currentDate.getFullYear()}-${currentDate.getMonth() + 1}-${currentDate.getDate()}`); // State to store the selected date
     const [numberOfMonths, setNumberOfMonths] = useState(3); // State to store the number of months to display
+
+    const [activeCard, setActiveCard] = useState<any>(null);
+
+    const [isDragging, setDragging] = useState(false);
 
     const currentMonth = new Date().toISOString().slice(0, 7);
 
     const [selectedMonth, setSelectedMonth] = useState(currentMonth);
 
-    const [dates, setDates] = useState(
-        getMonthsInView(currentDate, numberOfMonths)
-    ); // State to store the months to display
+    const [dates, setDates] = useState(getMonthsInView(currentDate, numberOfMonths)); // State to store the months to display
+
+    const [updateBt, { error, data }] = useMutation(UPDATE_BT);
 
     useEffect(() => {
-    if (editMode.isEditMode) {
-        props.setFilter("None");
-    }
+        if (editMode.isEditMode) {
+            props.setFilter("None");
+        }
     }, [editMode.isEditMode]);
-    
+
     const handleViewChange = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         // Get the form element from the event
         const form = event.target as HTMLFormElement;
 
         // Explicitly assert the type to HTMLInputElement and acces value
-        const dateInput = (
-            form.elements.namedItem("monthInput") as HTMLInputElement
-        )?.value;
-        const numberInput = (
-            form.elements.namedItem("numberInput") as HTMLInputElement
-        )?.value;
+        const dateInput = (form.elements.namedItem("monthInput") as HTMLInputElement)?.value;
+        const numberInput = (form.elements.namedItem("numberInput") as HTMLInputElement)?.value;
 
         setSelectedMonth(dateInput);
         setSelectedDate(dateInput);
@@ -126,11 +121,7 @@ function DisplayComponent(props: DisplayProps) {
         },
     });
 
-    const {
-        loading: loadingPendingBT,
-        error: errorPendingBT,
-        data: dataPendingBT,
-    } = useSubscription(GET_BT_PENDING_SUB);
+    const { loading: loadingPendingBT, error: errorPendingBT, data: dataPendingBT } = useSubscription(GET_BT_PENDING_SUB);
 
     if (loadingRigs) {
         return <p>Loading...</p>;
@@ -173,17 +164,11 @@ function DisplayComponent(props: DisplayProps) {
         setRigs(createRigs(numberOfRigs));
     }
 
-    
-
     //Making schedule BladeTaskCards
     let btCards: React.ReactNode[] = [];
     dataBT["AllBladeTasksInRangeSub"].forEach((bt: any) => {
         let btShown = false;
-        if (
-            bt.bladeProject.customer === props.filter ||
-            props.filter === "None" ||
-            editMode.isEditMode
-        ) {
+        if (bt.bladeProject.customer === props.filter || props.filter === "None" || editMode.isEditMode) {
             btShown = true;
         }
         let dateSplit = bt.startDate.split("-");
@@ -221,11 +206,7 @@ function DisplayComponent(props: DisplayProps) {
     let btCardsPending: React.ReactNode[] = [];
     dataPendingBT["AllBladeTasksPendingSub"].forEach((bt: any) => {
         let btShown = false;
-        if (
-            bt.bladeProject.customer === props.filter ||
-            props.filter === "None" ||
-            editMode.isEditMode
-        ) {
+        if (bt.bladeProject.customer === props.filter || props.filter === "None" || editMode.isEditMode) {
             btShown = true;
         }
 
@@ -248,6 +229,10 @@ function DisplayComponent(props: DisplayProps) {
         );
     });
 
+    let bladeTasksPendingHolder = new BladeTaskHolder(btCardsPending);
+
+    let bladeTasksHolder = new BladeTaskHolder(btCards);
+
     return (
         <div className="ScheduleContentContainer">
             <div className="ScheduleViewControl">
@@ -256,59 +241,112 @@ function DisplayComponent(props: DisplayProps) {
                         handleViewChange(e);
                     }}
                 >
-                    <input
-                        name="monthInput"
-                        type="month"
-                        defaultValue={selectedMonth}
-                    />
+                    <input name="monthInput" type="month" defaultValue={selectedMonth} />
                     <label htmlFor="numberInput" className="text-sm">
                         Months shown:
                     </label>
-                    <input
-                        className="numberMonthsInput"
-                        name="numberInput"
-                        type="number"
-                        defaultValue={numberOfMonths}
-                        min="2"
-                        max="24"
-                    />
+                    <input className="numberMonthsInput" name="numberInput" type="number" defaultValue={numberOfMonths} min="2" max="24" />
                     <input className="goButton" type="submit" value={"Go To"} />
                 </form>
             </div>
             {editMode.isEditMode ? (
-            <div className="ScheduleFilterAndMode">
-                <SwitchComponent setShowPasswordPrompt={props.setShowPasswordPrompt} />
-                {localStorage.getItem('token') && <StyledButton onClick={()=>{localStorage.removeItem('token'); window.location.reload();}}> Logout </StyledButton>}
-            </div>
+                <div className="ScheduleFilterAndMode">
+                    <SwitchComponent setShowPasswordPrompt={props.setShowPasswordPrompt} />
+                    {localStorage.getItem("token") && (
+                        <StyledButton
+                            onClick={() => {
+                                localStorage.removeItem("token");
+                                window.location.reload();
+                            }}
+                        >
+                            {" "}
+                            Logout{" "}
+                        </StyledButton>
+                    )}
+                </div>
             ) : (
-            <div className="ScheduleFilterAndMode">
-                <label>Filter:</label>
-                <select
-                    name="customerFilter"
-                    id="customerFilter"
-                    onChange={(e) => {
-                        props.setFilter(e.target.value);
-                    }}
-                >
-                    <option value="None">None</option>
-                    {customers.map((customer) => FilterCustomers(customer))}
-                </select>
-               <SwitchComponent setShowPasswordPrompt={props.setShowPasswordPrompt} />
-               {localStorage.getItem('token') && <StyledButton onClick={()=>{localStorage.removeItem('token'); window.location.reload();}}> Logout </StyledButton>}
-            </div>           
+                <div className="ScheduleFilterAndMode">
+                    <label>Filter:</label>
+                    <select
+                        name="customerFilter"
+                        id="customerFilter"
+                        onChange={(e) => {
+                            props.setFilter(e.target.value);
+                        }}
+                    >
+                        <option value="None">None</option>
+                        {customers.map((customer) => FilterCustomers(customer))}
+                    </select>
+                    <SwitchComponent setShowPasswordPrompt={props.setShowPasswordPrompt} />
+                    {localStorage.getItem("token") && (
+                        <StyledButton
+                            onClick={() => {
+                                localStorage.removeItem("token");
+                                window.location.reload();
+                            }}
+                        >
+                            {" "}
+                            Logout{" "}
+                        </StyledButton>
+                    )}
+                </div>
             )}
+            <DndContext // DndContext is used to enable drag and drop functionality
+                onDragStart={(event) => {
+                    handleDragStart(event, setDragging, setActiveCard);
+                }}
+                onDragEnd={(event) => {
+                    handleDragEnd(event, bladeTasksHolder, bladeTasksPendingHolder, setDragging, updateBt);
+                }}
+            >
+                {
+                    <PendingTasks
+                        bladeTaskHolder={bladeTasksPendingHolder}
+                        bladeTaskCards={bladeTasksPendingHolder.getBladeTasks()}
+                        numberOfRigs={numberOfRigs}
+                        showPasswordPrompt={props.showPasswordPrompt}
+                    />
+                }
 
-            <div className="ScheduleDisplay">
-                <CreateTestRigDivs rigs={rigs} />
-                <CreateTimelineField
-                    rigs={rigs}
-                    months={dates}
-                    btCards={btCards}
-                    btCardsPending={btCardsPending}
-                    showPasswordPrompt={props.showPasswordPrompt}
-                    isPendingTasksIncluded={true}
-                />
-            </div>
+                {createPortal(
+                    <DragOverlay>
+                        {activeCard && console.log("Drag overlay created for:", activeCard)}
+                        {activeCard && console.log("Drag overlay color", activeCard.duration)}
+
+                        {activeCard && (
+                            <BladeTaskCardOverlay
+                                duration={activeCard.duration}
+                                attachPeriod={activeCard.attachPeriod}
+                                detachPeriod={activeCard.detachPeriod}
+                                projectColor={activeCard.projectColor}
+                                projectName={activeCard.projectName}
+                                projectId={activeCard.projectId}
+                                customer={activeCard.customer}
+                                taskName={activeCard.taskName}
+                                id={activeCard.id}
+                                shown={activeCard.shown}
+                                enableDraggable={activeCard.enableDraggable}
+                            />
+                        )}
+                    </DragOverlay>,
+                    document.body
+                )}
+
+                <div className="ScheduleDisplay">
+                    <CreateTestRigDivs rigs={rigs} />
+                    <CreateTimelineField
+                        rigs={rigs}
+                        months={dates}
+                        btCards={btCards}
+                        btCardsPending={btCardsPending}
+                        showPasswordPrompt={props.showPasswordPrompt}
+                        isPendingTasksIncluded={true}
+                        setActiveCard={setActiveCard}
+                        isDragging={isDragging}
+                        setDragging={setDragging}
+                    />
+                </div>
+            </DndContext>
             {editMode.isEditMode ? <CreateAdditionalContent /> : null}
         </div>
     );
@@ -340,12 +378,7 @@ export function getMonthsInView(startDate: Date, numberOfMonths: number) {
     if (!isNaN(startDate.valueOf())) {
         for (let i = 1; i < numberOfMonths; i++) {
             let newDate = new Date(year, month + i);
-            let lastDay = getMonthLength(
-                capitalizeFirstLetter(
-                    newDate.toLocaleString("default", { month: "long" })
-                ),
-                startDate.getFullYear()
-            );
+            let lastDay = getMonthLength(capitalizeFirstLetter(newDate.toLocaleString("default", { month: "long" })), startDate.getFullYear());
             viewMonths.push(new Date(year, month + i, lastDay));
         }
     } else {
@@ -365,25 +398,12 @@ function getQueryDates(startDate: Date, endDate: Date) {
     let endDateDay = endDate.getDate();
 
     if (startDate === endDate) {
-        let monthLength = getMonthLength(
-            capitalizeFirstLetter(
-                startDate.toLocaleString("default", { month: "long" })
-            ),
-            startDate.getFullYear()
-        );
+        let monthLength = getMonthLength(capitalizeFirstLetter(startDate.toLocaleString("default", { month: "long" })), startDate.getFullYear());
         endDateDay = monthLength;
     }
 
-    startDateSTR = convertToQueryDate(
-        startDate.getFullYear(),
-        startDate.getMonth(),
-        startDateDay
-    );
-    endDateSTR = convertToQueryDate(
-        endDate.getFullYear(),
-        endDate.getMonth(),
-        endDateDay
-    );
+    startDateSTR = convertToQueryDate(startDate.getFullYear(), startDate.getMonth(), startDateDay);
+    endDateSTR = convertToQueryDate(endDate.getFullYear(), endDate.getMonth(), endDateDay);
     return { startDate: startDateSTR, endDate: endDateSTR };
 }
 
@@ -404,4 +424,234 @@ function FilterCustomers(customer: string) {
             {customer}
         </option>
     );
+}
+
+export function handleDragStart(
+    event: any,
+    setDragging: React.Dispatch<React.SetStateAction<boolean>>,
+    setActiveCard: React.Dispatch<React.SetStateAction<BladeTaskCardProps | null>>
+) {
+    const { active } = event;
+    if (active !== null) {
+        setDragging(true);
+    }
+
+    if (event.active.data.current.type === "BladeTaskCardProps") {
+        setActiveCard(event.active.data.current.props);
+    }
+}
+
+export function handleDragEnd(
+    event: any,
+    bladeTaskHolder: BladeTaskHolder,
+    bladeTaskHolderPending: BladeTaskHolder,
+    setDragging: React.Dispatch<React.SetStateAction<boolean>>,
+    updateBT: Function
+) {
+    const { active, over } = event; // active is the element being dragged, over is the element being dragged over
+    const updatedBladeTaskCards = bladeTaskHolder.getBladeTasks(); // Get the blade tasks from the BladeTaskHolder
+    const updatedBladeTaskCardsPending = bladeTaskHolderPending.getBladeTasks(); // Get the pending blade tasks from the BladeTaskHolder
+
+    const { statusBT, indexBT } = findBTIndex(
+        // Find the index of the blade task being dragged
+        updatedBladeTaskCards,
+        updatedBladeTaskCardsPending,
+        active
+    );
+
+    let draggedCard: React.ReactElement;
+
+    if (over !== null && indexBT !== -1) {
+        //get dragged card
+        if (statusBT === "scheduled") {
+            // If the blade task is scheduled, get the blade task from the scheduled blade tasks
+            draggedCard = updatedBladeTaskCards[indexBT] as React.ReactElement;
+        } else {
+            // Else, get the blade task from the pending blade tasks
+            draggedCard = updatedBladeTaskCardsPending[indexBT] as React.ReactElement;
+        }
+
+        //Moving to pending tasks
+        if (over.id === "droppablePendingTasksId") {
+            if (statusBT === "scheduled") {
+                updateBT({
+                    //update blade task in database
+                    variables: {
+                        id: draggedCard.props.id,
+                        startDate: "undefined",
+                        duration: draggedCard.props.duration,
+                        rig: 0,
+                    },
+                });
+
+                updatedBladeTaskCards.splice(indexBT, 1);
+
+                updatedBladeTaskCardsPending.push(
+                    // Add the blade task to the pending blade tasks
+                    <BladeTaskCard
+                        key={draggedCard.key}
+                        id={draggedCard.props.id}
+                        duration={draggedCard.props.duration}
+                        attachPeriod={draggedCard.props.attachPeriod}
+                        detachPeriod={draggedCard.props.detachPeriod}
+                        projectColor={draggedCard.props.projectColor}
+                        projectId={draggedCard.props.projectId}
+                        customer={draggedCard.props.customer}
+                        taskName={draggedCard.props.taskName}
+                        shown={draggedCard.props.shown}
+                        startDate={undefined}
+                        endDate={undefined}
+                        rig={undefined}
+                    />
+                );
+
+                bladeTaskHolder.setBladeTasks(updatedBladeTaskCards); // Update the blade tasks in the BladeTaskHolder
+                bladeTaskHolderPending.setBladeTasks(
+                    // Update the pending blade tasks in the BladeTaskHolder
+                    updatedBladeTaskCardsPending
+                );
+            } else {
+                console.log("a pending BT was moved to pending");
+            }
+        } else {
+            const overIdSlpit = over.id.split("-");
+            const overRig = parseInt(overIdSlpit[0].split(" ")[1]); // Get the rig number of the column being dragged over
+            const overDate = new Date(overIdSlpit[1], overIdSlpit[2], overIdSlpit[3]); // Get the date of the column being dragged over
+
+            // Check for overlap before updating the cards
+            const isOverlap = checkForOverlap(draggedCard, updatedBladeTaskCards, indexBT, overDate, overRig);
+
+            if (!isOverlap) {
+                let newEndDate = new Date(overDate);
+
+                newEndDate.setDate(
+                    // Set the end date of the blade task
+                    newEndDate.getDate() + draggedCard.props.duration - 1
+                );
+
+                updateBT({
+                    //update blade task in database
+                    variables: {
+                        id: draggedCard.props.id,
+                        startDate: formatDate(overDate),
+                        duration: draggedCard.props.duration,
+                        rig: overRig,
+                    },
+                });
+
+                if (statusBT === "scheduled") {
+                    updatedBladeTaskCards[indexBT] = // Update the blade task in the scheduled blade tasks
+                        (
+                            <BladeTaskCard
+                                key={draggedCard.key}
+                                id={draggedCard.props.id}
+                                duration={draggedCard.props.duration}
+                                projectColor={draggedCard.props.projectColor}
+                                projectId={draggedCard.props.projectId}
+                                customer={draggedCard.props.customer}
+                                taskName={draggedCard.props.taskName}
+                                startDate={new Date(overDate)}
+                                endDate={newEndDate}
+                                rig={overRig}
+                                attachPeriod={draggedCard.props.attachPeriod}
+                                detachPeriod={draggedCard.props.detachPeriod}
+                                enableDraggable={draggedCard.props.enableDraggable}
+                                shown={draggedCard.props.shown}
+                            />
+                        );
+
+                    bladeTaskHolder.setBladeTasks(updatedBladeTaskCards); // Update the scheduled blade tasks in the BladeTaskHolder
+                } else {
+                    updatedBladeTaskCardsPending.splice(indexBT, 1);
+
+                    updatedBladeTaskCards.push(
+                        // Add the blade task to the scheduled blade tasks
+                        <BladeTaskCard
+                            key={draggedCard.key}
+                            id={draggedCard.props.id}
+                            duration={draggedCard.props.duration}
+                            projectColor={draggedCard.props.projectColor}
+                            projectId={draggedCard.props.projectId}
+                            customer={draggedCard.props.customer}
+                            taskName={draggedCard.props.taskName}
+                            shown={draggedCard.props.shown}
+                            startDate={new Date(overDate)}
+                            endDate={newEndDate}
+                            attachPeriod={draggedCard.props.attachPeriod}
+                            detachPeriod={draggedCard.props.detachPeriod}
+                            enableDraggable={draggedCard.props.enableDraggable}
+                            rig={overRig}
+                        />
+                    );
+
+                    bladeTaskHolder.setBladeTasks(updatedBladeTaskCards); // Update the scheduled blade tasks in the BladeTaskHolder
+                    bladeTaskHolderPending.setBladeTasks(
+                        // Update the pending blade tasks in the BladeTaskHolder
+                        updatedBladeTaskCardsPending
+                    );
+                }
+            } else {
+                console.log("Overlap detected. drag opreation cancelled");
+            }
+            setDragging(false);
+        }
+    } else {
+        console.log("over is null");
+    }
+}
+
+export function findBTIndex(bladeTaskCards: any, bladeTaskCardsPending: any, activeComponent: any) {
+    for (let i: number = 0; i < bladeTaskCards.length; i++) {
+        if (bladeTaskCards[i] && activeComponent.id === bladeTaskCards[i].props.id) {
+            return { statusBT: "scheduled", indexBT: i };
+        }
+    }
+    for (let i: number = 0; i < bladeTaskCardsPending.length; i++) {
+        if (bladeTaskCardsPending[i] && activeComponent.id === bladeTaskCardsPending[i].props.id) {
+            return { statusBT: "pending", indexBT: i };
+        }
+    }
+    console.log("blade task not found");
+    return { statusBT: "not found", indexBT: -1 };
+}
+
+function checkForOverlap(draggedCard: any, bladeTaskCards: any, BTIndex: number, newStartDate: Date, overRig: number) {
+    //start-/end date for dragged card
+    const startDateA = newStartDate;
+    let endDateA = new Date(newStartDate);
+    endDateA.setDate(endDateA.getDate() + draggedCard.props.duration - 1);
+
+    for (let i: number = 0; i < bladeTaskCards.length; i++) {
+        //start-/end date for i'th card
+        const startDateB = bladeTaskCards[i].props.startDate;
+        const endDateB = bladeTaskCards[i].props.endDate;
+
+        if (i !== BTIndex) {
+            //skip comparison with itself
+            if (bladeTaskCards[i].props.rig === overRig || bladeTaskCards[i].props.projectId === draggedCard.props.projectId) {
+                // skip comparison with cards on different rigs unless it is from same project
+                //Overlaps are covered by two cases. A check is made for each case.
+                if (startDateA <= startDateB) {
+                    if (!(endDateA < startDateB)) {
+                        return true;
+                    }
+                } else if (startDateB < startDateA) {
+                    if (!(endDateB < startDateA)) {
+                        return true;
+                    }
+                }
+            }
+        }
+    }
+    return false;
+}
+
+function formatDate(date: Date) {
+    const year = date.getFullYear();
+    // getMonth() returns 0-11; add 1 to make it 1-12 and pad with '0' if needed
+    const month = (date.getMonth() + 1).toString().padStart(2, "0");
+    // getDate() returns 1-31; pad with '0' if needed
+    const day = date.getDate().toString().padStart(2, "0");
+
+    return `${year}-${month}-${day}`;
 }
