+++ src/components/Schedule/PendingTasks.css
@@ -5,10 +5,12 @@
     flex-direction: column;
     height: fit-content;
     min-height: 100px;
-    width: 1500px;
-    grid-row: 4;   
+    width: calc(90%);
+    grid-row: 3;   
     border: 2px solid rgba(0,0,0, 1);
     margin: 10px;
+    overflow-x: scroll;
+    grid-column: 1/3;
 }
 
 .whiteSpace{
@@ -16,7 +18,7 @@
     width: 1500px;
     grid-row: 4; 
     margin: 10px;
-    min-height: 105px;
+    min-height: 120px;
 }
 
 .pendingTasksContainerInner{
